added functionality to schedule mail

+++ Server/src/main/java/notiflow/server/Repository/EmailRepository.java
@@ -3,5 +3,9 @@ package notiflow.server.Repository;
 import notiflow.server.Entities.EmailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailRepository extends JpaRepository<EmailEntity, Integer> {
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface EmailRepository extends JpaRepository<EmailEntity, Long> {
+
 }
