+++ Server/src/main/java/notiflow/server/Repository/RecipientRepository.java
@@ -3,5 +3,5 @@ package notiflow.server.Repository;
 import notiflow.server.Entities.RecipientEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipientRepository extends JpaRepository<RecipientEntity, Integer> {
+public interface RecipientRepository extends JpaRepository<RecipientEntity, Long> {
 }
