+++ Server/src/main/java/notiflow/server/Repository/TemplateRepository.java
@@ -3,5 +3,5 @@ package notiflow.server.Repository;
 import notiflow.server.Entities.TemplateEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TemplateRepository extends JpaRepository<TemplateEntity, Integer> {
+public interface TemplateRepository extends JpaRepository<TemplateEntity, Long> {
 }
